finagle: Don't treat client discards as failures in FailureAccrualFactory
Problem

Currently, without customisation, the `FailureAccrualFactory` will treat client
discards as failures, which leads to endpoints being marked dead. Discards
are received server-side by protocols with control planes (e.g., Mux) and
translated into `ClientsDiscardedRequestException`s, which are then raised
on the `Future` being dispatched by the server.

Discards are sent by the client when it is no longer listening for the response
to the  request. Most commonly, this is because of a timeout. The "cause" of
timeout (i.e. the source of latency) could be upstream/ingress of the server,
the server's service itself, or a downstream/egress endpoint.

In first two cases (upstream or the server's service), the endpoint being
marked down by `FailureAccrual` is entirely not at fault, and should not be
marked down. In the third case, where it is the source of the latency, it should
be marked down, however, assuming the client talking to this endpoint is
configured with a reasonable timeout that is now being violated, the client's
`TimeoutFilter` will raise a `TimeoutException` and we will not need
to rely on a client discard from upstream.

Solution

Introduce a new FailureFlag, `ClientDiscard`, which is used for
`ClientDiscardedRequestExceptions` (we cannot simply match these because
they are typed to the protocol and the `FailureAccrualFactory` is
protocol-agnostic). Treat exceptions flagged thusly the same as we treat
`Ignorable` exceptions -- they should not count towards the endpoint being
marked down.

Also add stats for the failures received in `FailureAccrualFactory` -- both from
service acquisition and application.

Differential Revision: https://phabricator.twitter.biz/D1282833